Network.AWS.Firehose.CreateDeliveryStream

Description

Creates a delivery stream.

By default, you can create up to 20 delivery streams per region.

This is an asynchronous operation that immediately returns. The initial status of the delivery stream is CREATING . After the delivery stream is created, its status is ACTIVE and it now accepts data. Attempts to send data to a delivery stream that is not in the ACTIVE state cause an exception. To check the state of a delivery stream, use DescribeDeliveryStream .

A delivery stream is configured with a single destination: Amazon S3, Amazon Elasticsearch Service, or Amazon Redshift. You must specify only one of the following destination configuration parameters: ExtendedS3DestinationConfiguration , S3DestinationConfiguration , ElasticsearchDestinationConfiguration , or RedshiftDestinationConfiguration .

When you specify S3DestinationConfiguration , you can also provide the following optional values: BufferingHints , EncryptionConfiguration , and CompressionFormat . By default, if no BufferingHints value is provided, Firehose buffers data up to 5 MB or for 5 minutes, whichever condition is satisfied first. Note that BufferingHints is a hint, so there are some cases where the service cannot adhere to these conditions strictly; for example, record boundaries are such that the size is a little over or under the configured buffering size. By default, no encryption is performed. We strongly recommend that you enable encryption to ensure secure data storage in Amazon S3.

A few notes about Amazon Redshift as a destination:

An Amazon Redshift destination requires an S3 bucket as intermediate location, as Firehose first delivers data to S3 and then uses COPY syntax to load data into an Amazon Redshift table. This is specified in the RedshiftDestinationConfiguration.S3Configuration parameter.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationConfiguration.S3Configuration because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

We strongly recommend that you use the user name and password you provide exclusively with Firehose, and that the permissions for the account are restricted for Amazon Redshift INSERT permissions.

Firehose assumes the IAM role that is configured as part of the destination. The role should allow the Firehose principal to assume the role, and the role should have permissions that allows the service to deliver the data. For more information, see Amazon S3 Bucket Access in the Amazon Kinesis Firehose Developer Guide .

Creating a Request

createDeliveryStream

Arguments

:: Text
cdsDeliveryStreamName

-> CreateDeliveryStream

Creates a value of CreateDeliveryStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdsS3DestinationConfiguration - [Deprecated] The destination in Amazon S3. You can specify only one destination.

cdsRedshiftDestinationConfiguration - The destination in Amazon Redshift. You can specify only one destination.

cdsElasticsearchDestinationConfiguration - The destination in Amazon ES. You can specify only one destination.

cdsExtendedS3DestinationConfiguration - The destination in Amazon S3. You can specify only one destination.

cdsDeliveryStreamName - The name of the delivery stream. This name must be unique per AWS account in the same region. You can have multiple delivery streams with the same name if they are in different accounts or different regions.

data CreateDeliveryStream

See: createDeliveryStream smart constructor.

Request Lenses

cdsS3DestinationConfiguration :: Lens' CreateDeliveryStream (Maybe S3DestinationConfiguration)

Deprecated

The destination in Amazon S3. You can specify only one destination.

cdsRedshiftDestinationConfiguration :: Lens' CreateDeliveryStream (Maybe RedshiftDestinationConfiguration)

The destination in Amazon Redshift. You can specify only one destination.

cdsElasticsearchDestinationConfiguration :: Lens' CreateDeliveryStream (Maybe ElasticsearchDestinationConfiguration)

The destination in Amazon ES. You can specify only one destination.

cdsExtendedS3DestinationConfiguration :: Lens' CreateDeliveryStream (Maybe ExtendedS3DestinationConfiguration)

The destination in Amazon S3. You can specify only one destination.

cdsDeliveryStreamName :: Lens' CreateDeliveryStream Text

The name of the delivery stream. This name must be unique per AWS account in the same region. You can have multiple delivery streams with the same name if they are in different accounts or different regions.

Destructuring the Response

createDeliveryStreamResponse

Arguments

:: Int
cdsrsResponseStatus

-> CreateDeliveryStreamResponse

Creates a value of CreateDeliveryStreamResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdsrsDeliveryStreamARN - The ARN of the delivery stream.

cdsrsResponseStatus - -- | The response status code.

data CreateDeliveryStreamResponse

See: createDeliveryStreamResponse smart constructor.

Response Lenses

cdsrsDeliveryStreamARN :: Lens' CreateDeliveryStreamResponse (Maybe Text)

The ARN of the delivery stream.

cdsrsResponseStatus :: Lens' CreateDeliveryStreamResponse Int

- | The response status code.