Network.AWS.Firehose.DescribeDeliveryStream

Description

Describes the specified delivery stream and gets the status. For example, after your delivery stream is created, call DescribeDeliveryStream to see if the delivery stream is ACTIVE and therefore ready for data to be sent to it.

Creating a Request

describeDeliveryStream

Arguments

:: Text
ddsDeliveryStreamName

-> DescribeDeliveryStream

Creates a value of DescribeDeliveryStream with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddsExclusiveStartDestinationId - The ID of the destination to start returning the destination information. Currently Firehose supports one destination per delivery stream.

ddsLimit - The limit on the number of destinations to return. Currently, you can have one destination per delivery stream.

ddsDeliveryStreamName - The name of the delivery stream.

data DescribeDeliveryStream

See: describeDeliveryStream smart constructor.

Request Lenses

ddsExclusiveStartDestinationId :: Lens' DescribeDeliveryStream (Maybe Text)

The ID of the destination to start returning the destination information. Currently Firehose supports one destination per delivery stream.

ddsLimit :: Lens' DescribeDeliveryStream (Maybe Natural)

The limit on the number of destinations to return. Currently, you can have one destination per delivery stream.

ddsDeliveryStreamName :: Lens' DescribeDeliveryStream Text

The name of the delivery stream.

Destructuring the Response

describeDeliveryStreamResponse

Arguments

:: Int
ddsrsResponseStatus

-> DeliveryStreamDescription
ddsrsDeliveryStreamDescription

-> DescribeDeliveryStreamResponse

Creates a value of DescribeDeliveryStreamResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddsrsResponseStatus - -- | The response status code.

ddsrsDeliveryStreamDescription - Information about the delivery stream.

data DescribeDeliveryStreamResponse

See: describeDeliveryStreamResponse smart constructor.

Response Lenses

ddsrsResponseStatus :: Lens' DescribeDeliveryStreamResponse Int

- | The response status code.

ddsrsDeliveryStreamDescription :: Lens' DescribeDeliveryStreamResponse DeliveryStreamDescription

Information about the delivery stream.