Network.AWS.Firehose.PutRecord

Description

Writes a single data record into an Amazon Kinesis Firehose delivery stream. To write multiple data records into a delivery stream, use PutRecordBatch . Applications using these operations are referred to as producers.

By default, each delivery stream can take in up to 2,000 transactions per second, 5,000 records per second, or 5 MB per second. Note that if you use PutRecord and PutRecordBatch , the limits are an aggregate across these two operations for each delivery stream. For more information about limits and how to request an increase, see Amazon Kinesis Firehose Limits .

You must specify the name of the delivery stream and the data record when using PutRecord . The data record consists of a data blob that can be up to 1,000 KB in size, and any kind of data, for example, a segment from a log file, geographic location data, web site clickstream data, etc.

Firehose buffers records before delivering them to the destination. To disambiguate the data blobs at the destination, a common solution is to use delimiters in the data, such as a newline (n ) or some other character unique within the data. This allows the consumer application(s) to parse individual data items when reading the data from the destination.

The PutRecord operation returns a RecordId , which is a unique string assigned to each record. Producer applications can use this ID for purposes such as auditability and investigation.

If the PutRecord operation throws a ServiceUnavailableException , back off and retry. If the exception persists, it is possible that the throughput limits have been exceeded for the delivery stream.

Data records sent to Firehose are stored for 24 hours from the time they are added to a delivery stream as it attempts to send the records to the destination. If the destination is unreachable for more than 24 hours, the data is no longer available.

Creating a Request

putRecord

Arguments

:: Text
prDeliveryStreamName

-> Record
prRecord

-> PutRecord

Creates a value of PutRecord with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prDeliveryStreamName - The name of the delivery stream.

prRecord - The record.

data PutRecord

See: putRecord smart constructor.

Request Lenses

prDeliveryStreamName :: Lens' PutRecord Text

The name of the delivery stream.

prRecord :: Lens' PutRecord Record

The record.

Destructuring the Response

putRecordResponse

Arguments

:: Int
prrsResponseStatus

-> Text
prrsRecordId

-> PutRecordResponse

Creates a value of PutRecordResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prrsResponseStatus - -- | The response status code.

prrsRecordId - The ID of the record.

data PutRecordResponse

See: putRecordResponse smart constructor.

Response Lenses

prrsResponseStatus :: Lens' PutRecordResponse Int

- | The response status code.

prrsRecordId :: Lens' PutRecordResponse Text

The ID of the record.