Network.AWS.Firehose.PutRecordBatch

Description

Writes multiple data records into a delivery stream in a single call, which can achieve higher throughput per producer than when writing single records. To write single data records into a delivery stream, use PutRecord . Applications using these operations are referred to as producers.

By default, each delivery stream can take in up to 2,000 transactions per second, 5,000 records per second, or 5 MB per second. Note that if you use PutRecord and PutRecordBatch , the limits are an aggregate across these two operations for each delivery stream. For more information about limits, see Amazon Kinesis Firehose Limits .

Each PutRecordBatch request supports up to 500 records. Each record in the request can be as large as 1,000 KB (before 64-bit encoding), up to a limit of 4 MB for the entire request. These limits cannot be changed.

You must specify the name of the delivery stream and the data record when using PutRecord . The data record consists of a data blob that can be up to 1,000 KB in size, and any kind of data, for example, a segment from a log file, geographic location data, web site clickstream data, and so on.

Firehose buffers records before delivering them to the destination. To disambiguate the data blobs at the destination, a common solution is to use delimiters in the data, such as a newline (n ) or some other character unique within the data. This allows the consumer application(s) to parse individual data items when reading the data from the destination.

The PutRecordBatch response includes a count of failed records, FailedPutCount , and an array of responses, RequestResponses . Each entry in the RequestResponses array provides additional information about the processed record, and directly correlates with a record in the request array using the same ordering, from the top to the bottom. The response array always includes the same number of records as the request array. RequestResponses includes both successfully and unsuccessfully processed records. Firehose attempts to process all records in each PutRecordBatch request. A single record failure does not stop the processing of subsequent records.

A successfully processed record includes a RecordId value, which is unique for the record. An unsuccessfully processed record includes ErrorCode and ErrorMessage values. ErrorCode reflects the type of error, and is one of the following values: ServiceUnavailable or InternalFailure . ErrorMessage provides more detailed information about the error.

If there is an internal server error or a timeout, the write might have completed or it might have failed. If FailedPutCount is greater than 0, retry the request, resending only those records that might have failed processing. This minimizes the possible duplicate records and also reduces the total bytes sent (and corresponding charges). We recommend that you handle any duplicates at the destination.

If PutRecordBatch throws ServiceUnavailableException , back off and retry. If the exception persists, it is possible that the throughput limits have been exceeded for the delivery stream.

Data records sent to Firehose are stored for 24 hours from the time they are added to a delivery stream as it attempts to send the records to the destination. If the destination is unreachable for more than 24 hours, the data is no longer available.

Creating a Request

putRecordBatch

Arguments

:: Text
prbDeliveryStreamName

-> NonEmpty Record
prbRecords

-> PutRecordBatch

Creates a value of PutRecordBatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prbDeliveryStreamName - The name of the delivery stream.

prbRecords - One or more records.

data PutRecordBatch

See: putRecordBatch smart constructor.

Request Lenses

prbDeliveryStreamName :: Lens' PutRecordBatch Text

The name of the delivery stream.

prbRecords :: Lens' PutRecordBatch (NonEmpty Record)

One or more records.

Destructuring the Response

putRecordBatchResponse

Arguments

:: Int
prbrsResponseStatus

-> Natural
prbrsFailedPutCount

-> NonEmpty PutRecordBatchResponseEntry
prbrsRequestResponses

-> PutRecordBatchResponse

Creates a value of PutRecordBatchResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prbrsResponseStatus - -- | The response status code.

prbrsFailedPutCount - The number of records that might have failed processing.

prbrsRequestResponses - The results array. For each record, the index of the response element is the same as the index used in the request array.

data PutRecordBatchResponse

See: putRecordBatchResponse smart constructor.

Response Lenses

prbrsResponseStatus :: Lens' PutRecordBatchResponse Int

- | The response status code.

prbrsFailedPutCount :: Lens' PutRecordBatchResponse Natural

The number of records that might have failed processing.

prbrsRequestResponses :: Lens' PutRecordBatchResponse (NonEmpty PutRecordBatchResponseEntry)

The results array. For each record, the index of the response element is the same as the index used in the request array.