Network.AWS.Firehose.UpdateDestination

Description

Updates the specified destination of the specified delivery stream.

You can use this operation to change the destination type (for example, to replace the Amazon S3 destination with Amazon Redshift) or change the parameters associated with a destination (for example, to change the bucket name of the Amazon S3 destination). The update might not occur immediately. The target delivery stream remains active while the configurations are updated, so data writes to the delivery stream can continue during this process. The updated configurations are usually effective within a few minutes.

Note that switching between Amazon ES and other services is not supported. For an Amazon ES destination, you can only update to another Amazon ES destination.

If the destination type is the same, Firehose merges the configuration parameters specified with the destination configuration that already exists on the delivery stream. If any of the parameters are not specified in the call, the existing values are retained. For example, in the Amazon S3 destination, if EncryptionConfiguration is not specified then the existing EncryptionConfiguration is maintained on the destination.

If the destination type is not the same, for example, changing the destination from Amazon S3 to Amazon Redshift, Firehose does not merge any parameters. In this case, all parameters must be specified.

Firehose uses CurrentDeliveryStreamVersionId to avoid race conditions and conflicting merges. This is a required field, and the service updates the configuration only if the existing configuration has a version ID that matches. After the update is applied successfully, the version ID is updated, and can be retrieved using DescribeDeliveryStream . You should use the new version ID to set CurrentDeliveryStreamVersionId in the next call.

Creating a Request

updateDestination

Arguments

:: Text
udDeliveryStreamName

-> Text
udCurrentDeliveryStreamVersionId

-> Text
udDestinationId

-> UpdateDestination

Creates a value of UpdateDestination with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udS3DestinationUpdate - [Deprecated] Describes an update for a destination in Amazon S3.

udRedshiftDestinationUpdate - Describes an update for a destination in Amazon Redshift.

udElasticsearchDestinationUpdate - Describes an update for a destination in Amazon ES.

udExtendedS3DestinationUpdate - Describes an update for a destination in Amazon S3.

udDeliveryStreamName - The name of the delivery stream.

udCurrentDeliveryStreamVersionId - Obtain this value from the VersionId result of DeliveryStreamDescription . This value is required, and helps the service to perform conditional operations. For example, if there is a interleaving update and this value is null, then the update destination fails. After the update is successful, the VersionId value is updated. The service then performs a merge of the old configuration with the new configuration.

udDestinationId - The ID of the destination.

data UpdateDestination

See: updateDestination smart constructor.

Request Lenses

udS3DestinationUpdate :: Lens' UpdateDestination (Maybe S3DestinationUpdate)

Deprecated

Describes an update for a destination in Amazon S3.

udRedshiftDestinationUpdate :: Lens' UpdateDestination (Maybe RedshiftDestinationUpdate)

Describes an update for a destination in Amazon Redshift.

udElasticsearchDestinationUpdate :: Lens' UpdateDestination (Maybe ElasticsearchDestinationUpdate)

Describes an update for a destination in Amazon ES.

udExtendedS3DestinationUpdate :: Lens' UpdateDestination (Maybe ExtendedS3DestinationUpdate)

Describes an update for a destination in Amazon S3.

udDeliveryStreamName :: Lens' UpdateDestination Text

The name of the delivery stream.

udCurrentDeliveryStreamVersionId :: Lens' UpdateDestination Text

Obtain this value from the VersionId result of DeliveryStreamDescription . This value is required, and helps the service to perform conditional operations. For example, if there is a interleaving update and this value is null, then the update destination fails. After the update is successful, the VersionId value is updated. The service then performs a merge of the old configuration with the new configuration.

udDestinationId :: Lens' UpdateDestination Text

The ID of the destination.

Destructuring the Response

updateDestinationResponse

Arguments

:: Int
udrsResponseStatus

-> UpdateDestinationResponse

Creates a value of UpdateDestinationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udrsResponseStatus - -- | The response status code.

data UpdateDestinationResponse

See: updateDestinationResponse smart constructor.

Response Lenses

udrsResponseStatus :: Lens' UpdateDestinationResponse Int

- | The response status code.